adapter/requirements.txt
paho-mqtt
influxdb
python-dotenv
influxdb_client
Faker